requirements.txt
backports.weakref==1.0rc1
bleach==1.5.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.1.1
grpcio==1.4.0
html5lib==0.9999999
Markdown==2.6.9
mock==2.0.0
numpy==1.13.1
pbr==3.1.1
pkg-resources==0.0.0
protobuf==3.4.0
six==1.10.0
tensorflow==1.3.0
tensorflow-serving-api==1.0.0
tensorflow-tensorboard==0.1.6
Werkzeug==0.12.2
